// types/configuracion.ts
export interface ConfiguracionSistema {
  nombreSistema: string;
  nombreInstitucion: string;
  descripcion: string;
  correoContacto: string;
  telefonoContacto: string;
  sitioWeb: string;
  mostrarNoticias: boolean;
  mostrarDonaciones: boolean;
  mostrarMapa: boolean;
  modoMantenimiento: boolean;
}